adding search feature : user can search other users and posts

+++ app/lib/functions/user/post/deletePost.tsx
@@ -14,20 +14,45 @@ async function deletePostAction(id: number) {
     const { data: list, error } = await supabase.storage
       .from("posts")
       .list(`public/${user.id}/${id}`);
-    if (list) {
+    if (error) {
+      throw new Error(
+        `message : ${error.message},cause : ${error.cause},name : ${error.name}`
+      );
+    }
+    if (list && list?.length > 0) {
       const filesToRemove = list.map(
         (x) => `public/${user.id}/${id}/${x.name}`
       );
+
       const { data, error } = await supabase.storage
         .from("posts")
         .remove(filesToRemove);
+      if (error) {
+        throw new Error(
+          `message : ${error.message},cause : ${error.cause},name : ${error.name}`
+        );
+      }
       const { data: post_delete, error: error_delete } = await supabase
         .from("posts")
         .delete()
         .eq("id", id);
+      if (error_delete) {
+        throw new Error(
+          `message : ${error_delete.message},details : ${error_delete.details},hint : ${error_delete.hint}`
+        );
+      }
     }
-    revalidatePath("/home");
+    const { data: post_delete, error: error_delete } = await supabase
+      .from("posts")
+      .delete()
+      .eq("id", id);
+    if (error_delete) {
+      throw new Error(
+        `message : ${error_delete.message},details : ${error_delete.details},hint : ${error_delete.hint}`
+      );
+    }
   }
+  revalidatePath("/home");
 }
 
 export default deletePostAction;
